requirements.txt
streamlit
torch
transformers
diffusers>=0.14.0
Pillow
scipy
accelerate